Finish HTML part
This basic website consist of a homepage and three recipes page. In the homepage there are three links that if pressed will bring to one of the recipe page that consist of end product image, description of the food, ingredients, step by step instructions, and a link that will bring you to back to the homepage.